+++ .gitignore
@@ -1,1 +1,3 @@
 ./old_bot
+.idea/
+__pycache__
