+++ application/frontend/frontend.py
@@ -3,6 +3,6 @@ from fastapi import FastAPI
 app = FastAPI()
 
 
-@app.get("/")
+@app.get("/users/count")
 async def user_count():
-    return
+    return dict(count=0);
